--- components/consent-mgt-server-configs/org.wso2.carbon.identity.consent.mgt.server.configs/pom.xml
@@ -1,99 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- ~ Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org).
- ~
- ~ WSO2 LLC. licenses this file to you under the Apache License,
- ~ Version 2.0 (the "License"); you may not use this file except
- ~ in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~    http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing,
- ~ software distributed under the License is distributed on an
- ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- ~ KIND, either express or implied.  See the License for the
- ~ specific language governing permissions and limitations
- ~ under the License.
- -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <groupId>org.wso2.carbon.identity.framework</groupId>
-        <artifactId>consent-mgt-server-configs</artifactId>
-        <version>5.25.173-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.identity.consent.mgt.server.configs</artifactId>
-    <packaging>bundle</packaging>
-    <name>WSO2 Carbon Identity- Consent Mgt Server Configs</name>
-    <description>A custom wso2 products or solution</description>
-    <url>http://wso2.org</url>
-    <dependencies>
-        <dependency>
-            <groupId>org.wso2.eclipse.osgi</groupId>
-            <artifactId>org.eclipse.osgi.services</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.felix</groupId>
-            <artifactId>org.apache.felix.scr.ds-annotations</artifactId>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon.identity.framework</groupId>
-            <artifactId>org.wso2.carbon.identity.configuration.mgt.core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.testng</groupId>
-            <artifactId>testng</artifactId>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.mockito</groupId>
-            <artifactId>mockito-core</artifactId>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.felix</groupId>
-                <artifactId>maven-bundle-plugin</artifactId>
-                <extensions>true</extensions>
-                <configuration>
-                    <instructions>
-                        <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Name>${project.artifactId}</Bundle-Name>
-                        <Axis2Module>${project.artifactId}-${project.version}</Axis2Module>
-                        <Private-Package>
-                            org.wso2.carbon.identity.consent.mgt.server.configs.internal,
-                        </Private-Package>
-                        <Import-Package>
-                            org.wso2.carbon.identity.configuration.mgt.core.*;
-                            version="${carbon.identity.package.import.version.range}",
-                            org.apache.commons.logging; version="${import.package.version.commons.logging}",
-                            org.osgi.framework; version="${osgi.framework.imp.pkg.version.range}",
-                            org.osgi.service.component; version="${osgi.service.component.imp.pkg.version.range}",
-                        </Import-Package>
-                        <Export-Package>
-                            !org.wso2.carbon.identity.consent.mgt.server.configs.internal,
-                            org.wso2.carbon.identity.consent.mgt.server.configs.*;
-                            version="${carbon.identity.package.export.version}",
-                        </Export-Package>
-                    </instructions>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>findbugs-maven-plugin</artifactId>
-                <configuration>
-                    <threshold>High</threshold>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-
-</project>
