--- components/consent-mgt-server-configs/org.wso2.carbon.identity.consent.mgt.server.configs/src/main/java/org/wso2/carbon/identity/consent/mgt/server/configs/exceptions/ConsentMgtServerConfigsClientException.java
@@ -1,70 +1,0 @@
-/*
- * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com).
- *
- * WSO2 LLC. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.wso2.carbon.identity.consent.mgt.server.configs.exceptions;
-
-/**
- * Consent Management Server Configs related server exceptions.
- */
-public class ConsentMgtServerConfigsClientException extends ConsentMgtServerConfigsException {
-
-    /**
-     * Constructor with error code and message.
-     *
-     * @param errorCode Error Code.
-     * @param message   Message.
-     */
-    public ConsentMgtServerConfigsClientException(String errorCode, String message) {
-
-        super(errorCode, message);
-    }
-
-    /**
-     * Constructor with error code, message and description.
-     *
-     * @param errorCode     Error Code.
-     * @param message       Message.
-     * @param description   Description.
-     */
-    public ConsentMgtServerConfigsClientException(String errorCode, String message, String description) {
-
-        super(errorCode, message, description);
-    }
-
-    /**
-     * Constructor with error code, message and cause.
-     *
-     * @param errorCode Error Code.
-     * @param message   Error message.
-     * @param cause     If any error occurred when accessing the tenant.
-     */
-    public ConsentMgtServerConfigsClientException(String errorCode, String message, Throwable cause) {
-
-        super(errorCode, message, cause);
-    }
-
-    /**
-     * Constructor with cause.
-     *
-     * @param cause If any error occurred when accessing the tenant.
-     */
-    public ConsentMgtServerConfigsClientException(Throwable cause) {
-
-        super(cause);
-    }
-}
