--- components/consent-mgt-server-configs/org.wso2.carbon.identity.consent.mgt.server.configs/src/test/resources/testing.xml
@@ -1,27 +1,0 @@
-<!--
-  ~ Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com).
-  ~
-  ~ WSO2 LLC. licenses this file to you under the Apache License,
-  ~ Version 2.0 (the "License"); you may not use this file except
-  ~ in compliance with the License.
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing,
-  ~ software distributed under the License is distributed on an
-  ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-  ~ KIND, either express or implied.  See the License for the
-  ~ specific language governing permissions and limitations
-  ~ under the License.
-  -->
-
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
-
-<suite name="org.wso2.carbon.identity.consent.mgt.server.configs.test">
-    <test name="org.wso2.carbon.identity.consent.mgt.server.configs.test" preserve-order="false" parallel="false">
-        <!--<parameter name="log-level" value="debug"/>-->
-        <classes>
-            <class name="org.wso2.carbon.identity.consent.mgt.server.configs.test.ConsentManagementServerConfigsServiceTest"/>
-        </classes>
-    </test>
-</suite>
