--- components/consent-mgt-server-configs/pom.xml
@@ -1,38 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
- ~ Copyright (c) 2023, WSO2 LLC. (http://www.wso2.org).
- ~
- ~ WSO2 LLC. licenses this file to you under the Apache License,
- ~ Version 2.0 (the "License"); you may not use this file except
- ~ in compliance with the License.
- ~ You may obtain a copy of the License at
- ~
- ~    http://www.apache.org/licenses/LICENSE-2.0
- ~
- ~ Unless required by applicable law or agreed to in writing,
- ~ software distributed under the License is distributed on an
- ~ "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- ~ KIND, either express or implied.  See the License for the
- ~ specific language governing permissions and limitations
- ~ under the License.
- -->
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-
-    <parent>
-        <groupId>org.wso2.carbon.identity.framework</groupId>
-        <artifactId>identity-framework</artifactId>
-        <version>5.25.173-SNAPSHOT</version>
-        <relativePath>../../pom.xml</relativePath>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>consent-mgt-server-configs</artifactId>
-    <packaging>pom</packaging>
-    <name>WSO2 Carbon - Identity Consent Mgt Server Configs Module</name>
-    <url>http://wso2.org</url>
-
-    <modules>
-        <module>org.wso2.carbon.identity.consent.mgt.server.configs</module>
-    </modules>
-
-</project>
