+++ components/consent-server-configs-mgt/org.wso2.carbon.identity.consent.server.configs.mgt/src/main/java/org/wso2/carbon/identity/consent/server/configs/mgt/services/ConsentServerConfigsManagementService.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2023, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package org.wso2.carbon.identity.consent.server.configs.mgt.services;
+
+import org.wso2.carbon.identity.consent.server.configs.mgt.exceptions.ConsentServerConfigsMgtException;
+
+/**
+ * Interface for Consent Server Configs Management Service.
+ */
+public interface ConsentServerConfigsManagementService {
+
+   /**
+    * Method to get external consent page url.
+    *
+    * @param tenantDomain  Tenant domain.
+    * @return External consent page url.
+    * @throws ConsentServerConfigsMgtException If an error occurred in getting the consent url.
+    */
+    String getExternalConsentPageUrl(String tenantDomain) throws ConsentServerConfigsMgtException;
+
+}
