+++ pom.xml
@@ -69,7 +69,7 @@
         <module>components/secret-mgt</module>
         <module>components/central-logger</module>
         <module>components/input-validation-mgt</module>
-        <module>components/consent-mgt-server-configs</module>
+        <module>components/consent-server-configs-mgt</module>
         <module>features/security-mgt</module>
         <module>features/claim-mgt</module>
         <module>features/identity-core</module>
@@ -98,7 +98,7 @@
         <module>features/secret-mgt</module>
         <module>features/central-logger</module>
         <module>features/input-validation-mgt</module>
-        <module>features/consent-mgt-server-configs</module>
+        <module>features/consent-server-configs-mgt</module>
     </modules>
 
 
@@ -291,7 +291,7 @@
 
             <dependency>
                 <groupId>org.wso2.carbon.identity.framework</groupId>
-                <artifactId>org.wso2.carbon.identity.consent.mgt.server.configs</artifactId>
+                <artifactId>org.wso2.carbon.identity.consent.server.configs.mgt</artifactId>
                 <version>${project.version}</version>
             </dependency>
